+++ client/src/routes/+layout.svelte
@@ -1,6 +1,6 @@
 <script>
 	import '../app.css';
-	import Navigation from './components/Navigation.svelte';
+	import Navigation from '../components/Navigation.svelte';
 </script>
 
 <Navigation />
